[telemetry] Make the Iris node agent own hardware metrics
