+++ lib/iris/src/iris/cluster/composer.py
@@ -66,7 +66,6 @@ def make_task_backend(
     task_stats_table: Table | None = None,
     task_event_table: Table | None = None,
     profile_table: Table | None = None,
-    worker_stats_table: Table | None = None,
     autoscaler: Autoscaler | None = None,
     transition_reader: TransitionReader | None = None,
 ) -> TaskBackend:
@@ -153,7 +152,6 @@ def make_task_backend(
             task_stats_table=task_stats_table,
             task_event_table=task_event_table,
             profile_table=profile_table,
-            worker_stats_table=worker_stats_table,
             transition_reader=transition_reader,
         )
     if which == "worker_provider":
@@ -232,7 +230,6 @@ def make_backend(
             task_stats_table=log_stack.task_stats_table,
             task_event_table=log_stack.task_event_table,
             profile_table=log_stack.profile_table,
-            worker_stats_table=log_stack.worker_stats_table,
             transition_reader=DbTransitionReader(db),
         )
         logger.info("Backend created: %s", type(provider).__name__)
