+++ lib/iris/src/iris/cluster/node_agent/__init__.py
@@ -1,0 +1,6 @@
+# Copyright The Marin Authors
+# SPDX-License-Identifier: Apache-2.0
+
+"""Iris-owned physical node telemetry."""
+
+SERVICE_NAME = "iris-node-agent"
